+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # dev
 
+* nikita now depends on node.js version >= 8.12 and npm >= 6.4
 * improved styling and structure of all boilerplate components
 * added more boilerplate components like header, footer, button and form elements
 * added `MediaQuery.js` lib as an adapter to foundation breakpoints
 * added twig.js functions `html_attr` and `html_classes`
+* updated all dependencies to newest version
+* updated nikita eslint and stylelint config
 * fixed grunt task `analyze-js`
 * renamed `App.js` to PascalCase
 * fixed generator version compare bug