+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 * improved styling and structure of all boilerplate components
 * moved color declarations to `scss/variables/_foundation-settings.scss` and removed file `scss/variables/_color.scss`
 * added more boilerplate components like header, footer, button and form elements
+* added `react-helmet-async` lib to modify html document head
 * added `MediaQuery.js` lib as an adapter to foundation breakpoints
 * added `KeyboardFocus.js` lib to restrict focus styles to keyboard navigation
 * added twig.js functions `html_attr` and `html_classes`
-* changed twig namespaces to  at-notation
+* changed twig namespaces to  @-notation
 * updated all dependencies to newest version
 * updated nikita eslint and stylelint config
 * fixed grunt task `analyze-js`