+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * added nikita version to npm generator script target
 * internal: fixed generator version compare bug
 * internal: added method to copy complete template directory to generator
+* added grunt task to check licenses used in the project
 
 # 5.2.0 (2019/05/08)