+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# dev
+# 6.0.0 (2020/04/29)
 
 * nikita now depends on node.js version >= 8.12 and npm >= 6.4
 * improved styling and structure of all boilerplate components